@@ -11,52 +11,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.1",
-    "@angular/cdk": "^13.1.1",
-    "@angular/common": "^13.1.1",
-    "@angular/compiler": "^13.1.1",
-    "@angular/core": "^13.1.1",
-    "@angular/flex-layout": "^13.0.0-beta.36",
-    "@angular/forms": "^13.1.1",
-    "@angular/material": "^13.1.1",
-    "@angular/platform-browser": "^13.1.1",
-    "@angular/platform-browser-dynamic": "^13.1.1",
-    "@angular/router": "^13.1.1",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
+    "@angular/animations": "^14.0.5",
+    "@angular/cdk": "^14.0.4",
+    "@angular/common": "^14.0.5",
+    "@angular/compiler": "^14.0.5",
+    "@angular/core": "^14.0.5",
+    "@angular/flex-layout": "^14.0.0-beta.40",
+    "@angular/forms": "^14.0.5",
+    "@angular/material": "^14.0.4",
+    "@angular/platform-browser": "^14.0.5",
+    "@angular/platform-browser-dynamic": "^14.0.5",
+    "@angular/router": "^14.0.5",
+    "@ng-bootstrap/ng-bootstrap": "^12.1.2",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@types/greensock": "^1.15.32",
     "bootstrap": "^5.1.3",
-    "core-js": "^3.20.2",
-    "gsap": "^3.9.1",
+    "core-js": "^3.23.3",
+    "gsap": "^3.10.4",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.6.7",
     "rxjs-compat": "^6.6.7",
     "tslib": "^2.1.0",
-    "zone.js": "^0.11.4"
+    "zone.js": "^0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.2",
-    "@angular/cli": "^13.1.2",
-    "@angular/compiler-cli": "^13.1.1",
-    "@angular/language-service": "^13.1.1",
-    "@types/jasmine": "^3.10.3",
+    "@angular-devkit/build-angular": "^14.0.5",
+    "@angular/cli": "^14.0.5",
+    "@angular/compiler-cli": "^14.0.5",
+    "@angular/language-service": "^14.0.5",
+    "@types/jasmine": "^4.0.3",
     "@types/jasminewd2": "^2.0.10",
-    "@types/node": "^17.0.8",
+    "@types/node": "^18.0.3",
     "codelyzer": "^6.0.2",
-    "jasmine-core": "^4.0.0",
+    "jasmine-core": "^4.2.0",
     "jasmine-spec-reporter": "^7.0.0",
-    "karma": "^6.3.10",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage": "^2.1.0",
+    "karma": "^6.4.0",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage": "^2.2.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
-    "karma-jasmine": "^4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
     "protractor": "^7.0.0",
     "ts-lint": "^4.5.1",
-    "ts-node": "^10.4.0",
+    "ts-node": "^10.8.2",
     "tslint": "^6.1.3",
-    "typescript": "^4.5.4",
+    "typescript": "^4.7.4",
     "webpack-bundle-analyzer": "^4.5.0"
   }
 }
